Fixes an issue reader Illo de' Illis noticed on the budget. Decided to remove having loan repayment being factored in a budget, making repayments more of a "surprise" factor.
The main reason for this is because of the way the loan periods are structured,
factoring in repayments into the Persons budget will cause a oscillatory price system before settling into equilibrium. The oscillations can be thought of as the results of shortrun costs coming together

Also, now that the simulation is a lot simpler, a more precise price-stepping is used. Instead of stepping $100 10 times, $100 is now stepped 200 times, making equilibrium prices be accurate to a $0.50 level